Determine what `WordBank` to pull words from
Point at a given `WordBank` based on a `SessionConfig` passed to the `Session` class constructor.

An alternative was to wrap the word banks in an array, and then index that array with a `SessionConfig::Difficulty` - this approach is not only less readable, but also allows the word bank to change in the middle of a `Session` depending on this index value (which is not the expected behaviour).

While the `m_wordBank` member is only a pointer to `const` (meaning its address can change), it *should* only be assigned in the `Session` class constructor, whereas this index value could technically be anything.