Make `SessionConfig::Difficulty` an unscoped enum
Initially `SessionConfig::Difficulty` wasn't intended to be used to index containers as much as is. This change makes it easier to do so, and allows the enumerators to be accessed directly.

While the type safety provided by having it be a scoped enum is useful, the added verbosity is unpleasant.