Move `static` members out of `Session` class definition
In the event that the file path passed to an `Audio::Track` doesn't exist an exception is thrown. While the correct file path can be passed directly into the `static` `Audio::Track` members in the `Session` class definition, any tests that use `Session` no longer pass, and crash with non-descriptive errors.

This approach allows tests to pass in empty audio tracks to a `Session`, and is more in line with how the main menu and leaderboard implement audio.