Refactor on-screen word generation and deletion
In `Session::getRandomWord()` the starting string was always initially empty, meaning that the first iteration of the following loop would always result in an assignment.

In `Session::handleInput()` the contents of the input component was being checked against all current on-screen words twice; once to ensure that it matched an on-screen word, and a second time to actually locate and erase it. This change combines both these steps.