Fix `UNIX` based string comparison bug
In web builds, both the words stored in a given `WordBank`, and the `Word::text` members of dynamically allocated `Word` objects have an appended return carriage character before their null termination character.

This not only increases the length of these strings by one, but also results in unexpected behaviour when the `operator==` is used to compare the player's input string with the on-screen words.

This seems to be because Windows represents a newline as `CR + LF` ("\r\n"), whereas UNIX - which Emscripten emulates - represents it as `LF` ("\n"). Since the word banks are Windows text files, when they are read from in a web build, the "\r" is left behind.

This fix just checks for both versions.