Refactor `miniaudio.h` Web Audio `ma_device` initialisation
When building to WebAssembly using Emscripten with `PTHREADS` and the `PROXY_TO_PTHREAD` flags enabled (which are necessary for `FTXUI` to run smoothly) `miniaudio` fails when initialising/uninitialising a `ma_device`.

This seems to be due to `PROXY_TO_PTHREAD` which moves the `main()` function into a pthread, meaning that `miniaudio`, which uses inline JavaScript via `EM_ASM()` to access DOM elements like `window` fails as these elements can only be interacted with in the main thread.

The fix for this is replacing inline JavaScript calls made via `EM_ASM()` with their main thread equivalents like `MAIN_THREAD_EM_ASM()`.